requirements.txt
# specific for endorse-experiment
h5py
scoop       # communication library
deap        # gnetic optimization
salib       # Sobol indices
PyPDF2      # PDF report compilation
img2pdf     # png to pdf
cartopy     #  For stereographic projection
#tables      # PyTables
#zarr        # alternative to H5PY
#pandas
#xarray


# endorse
pyvista
pyyaml
pyyaml-include
matplotlib
scipy
seaborn
pytest
tox
tox-gh-actions
importlib-resources
pandas
attrs

# bayes
statsmodels
ruamel.yaml     # ruamel keeps comments and order when loading and dumping yaml around

